Verteilung der TestKlassen in verschiedenen Ordern
Die TestKlassen werden in verschiedenen Ordnern verteilt, um die Analyse zu erleichtern. Jetzt ist es einfach zu wissen, wo die Test für Fallen oder für Monster sind.
