+++ game/test/FallenTest/TeleportpadsTest.java
@@ -1,53 +1,70 @@
-/*package FallenTest;
+package FallenTest;
 
 import ecs.components.AnimationComponent;
+import ecs.components.PositionComponent;
 import ecs.entities.Entity;
 import ecs.entities.Teleportpads;
 import ecs.entities.Teleportsystem;
+import level.elements.tile.Tile;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
-import static org.powermock.api.mockito.PowerMockito.verifyNew;
-import static org.powermock.api.mockito.PowerMockito.whenNew;
 
+/**
+ * Testklasse für die Klasse "Teleportpads".
+ */
 public class TeleportpadsTest {
 
+    private Teleportsystem mockTeleportsystem;
+
+    @BeforeEach
+    public void setup() {
+        mockTeleportsystem = mock(Teleportsystem.class);
+    }
+
+    /**
+     * Testet den Standardwert der "Usages" in der Klasse "Teleportpads".
+     */
     @Test
     public void testDefaultUsages() {
-        Teleportsystem mockTeleportSystem = mock(Teleportsystem.class);
-        Teleportpads teleportpads = new Teleportpads(3, mockTeleportSystem);
-        assertEquals(3, teleportpads.getUsages());
+        Teleportpads pads = new Teleportpads(10000, mockTeleportsystem);
+        assertEquals(10000, pads.getUsages());
     }
 
+    /**
+     * Testet die "Usages" in der Klasse "Teleportpads" mit einem benutzerdefinierten Wert.
+     */
     @Test
-    public void testAnimationComponentSetup() throws Exception {
-        AnimationComponent mockAnimationComponent = mock(AnimationComponent.class);
-        whenNew(AnimationComponent.class).withAnyArguments().thenReturn(mockAnimationComponent);
-
-        Teleportpads teleportpads = new Teleportpads(3, mock(Teleportsystem.class));
-
-        verifyNew(AnimationComponent.class).withArguments(eq(teleportpads), any(), any());
+    public void testCustomUsages() {
+        Teleportpads pads = new Teleportpads(500, mockTeleportsystem);
+        assertEquals(500, pads.getUsages());
     }
 
+    /**
+     * Testet das Vorhandensein einer PositionComponent in der Klasse "Teleportpads".
+     */
     @Test
-    public void testAniused() {
-        Teleportsystem mockTeleportSystem = mock(Teleportsystem.class);
-        Teleportpads teleportpads = new Teleportpads(3, mockTeleportSystem);
+    public void testPositionComponent() {
+        Teleportpads pads = new Teleportpads(10000, mockTeleportsystem);
+        assertNotNull(pads.getComponent(PositionComponent.class));
+    }
 
+    /**
+     * Testet die Animation beim Verlassen einer Kollision.
+     */
+    @Test
+    public void testCollisionLeaveAnimation() {
+        Teleportpads pads = new Teleportpads(10000, mockTeleportsystem);
         Entity mockEntity = mock(Entity.class);
         AnimationComponent mockAnimationComponent = mock(AnimationComponent.class);
-        when(mockEntity.getComponent(AnimationComponent.class)).thenReturn(Optional.of(mockAnimationComponent));
+        when(mockEntity.getComponent(AnimationComponent.class)).thenReturn(Optional.ofNullable(mockAnimationComponent));
 
-        teleportpads.Aniused(mockEntity);
+        pads.onCollisionleave(pads, mockEntity, Tile.Direction.W); //W = UP
 
-        verify(mockTeleportSystem).usedPad(teleportpads);
         verify(mockAnimationComponent).setCurrentAnimation(any());
     }
-
-    // Weitere Testfälle für andere Methoden der Klasse Teleportpads
-
 }
-*/
